refactor: input-item 컴포넌트 TypeScript로 마이그레이션
