+++ src/constants/formFields.ts
@@ -29,7 +29,8 @@ type TextareaField = BaseField & {
 
 type FormField = BaseField | SelectField | TextareaField;
 
-type FieldGroup = {
+export type FieldGroup = {
+  fields: Record<string, FormField>;
   create: (info: FormField) => HTMLElement;
 };
 
@@ -41,46 +42,50 @@ type FormFields = {
 
 export const FORM_FIELDS: FormFields = Object.freeze({
   INPUTS: Object.freeze({
-    name: {
-      label: "이름",
-      attribute: {
-        required: true,
-        id: "name",
-        name: "name",
-        type: "text",
-        maxlength: 30,
-        placeholder: "음식점 이름을 입력해주세요.",
+    fields: {
+      name: {
+        label: "이름",
+        attribute: {
+          required: true,
+          id: "name",
+          name: "name",
+          type: "text",
+          maxlength: 30,
+          placeholder: "음식점 이름을 입력해주세요.",
+        },
       },
-    },
-    link: {
-      label: "참고 링크",
-      attribute: {
-        id: "link",
-        name: "link",
-        type: "text",
-        maxlength: 100,
-        placeholder: "https://www.woowacourse.io/",
+      link: {
+        label: "참고 링크",
+        attribute: {
+          id: "link",
+          name: "link",
+          type: "text",
+          maxlength: 100,
+          placeholder: "https://www.woowacourse.io/",
+        },
       },
     },
     create: (info: FormField) => $input(info),
   }),
   SELECTS: Object.freeze({
-    category: {
-      label: "카테고리",
-      options: categoryOptions,
-      attribute: {
-        required: true,
-        id: "category",
-        name: "category",
+    fields: {
+      category: {
+        label: "카테고리",
+        options: categoryOptions,
+        attribute: {
+          required: true,
+          id: "category",
+          name: "category",
+        },
       },
-    },
-    distance: {
-      label: "거리(도보 이동 시간)",
-      options: distanceOptions,
-      attribute: {
-        required: true,
-        id: "distance",
-        name: "distance",
+      distance: {
+        label: "거리(도보 이동 시간)",
+        options: distanceOptions,
+        attribute: {
+          required: true,
+          id: "distance",
+          name: "distance",
+        },
       },
     },
     create: (info: FormField) => {
@@ -89,16 +94,18 @@ export const FORM_FIELDS: FormFields = Object.freeze({
     },
   }),
   TEXTAREAS: Object.freeze({
-    description: {
-      label: "설명",
-      helperText: "메뉴 등 추가 정보를 입력해 주세요.",
-      attribute: {
-        id: "description",
-        name: "description",
-        cols: "30",
-        rows: "5",
-        maxlength: 200,
-        placeholder: "너무 맛있는데 너무 매워서 배가 아파요,,,",
+    fields: {
+      description: {
+        label: "설명",
+        helperText: "메뉴 등 추가 정보를 입력해 주세요.",
+        attribute: {
+          id: "description",
+          name: "description",
+          cols: "30",
+          rows: "5",
+          maxlength: 200,
+          placeholder: "너무 맛있는데 너무 매워서 배가 아파요,,,",
+        },
       },
     },
     create: (info: FormField) => $textarea(info),
